# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: Construction Enterprise Operating Platform API
  version: "0.8.1"
  description: Unified enterprise operating platform for construction company governance, business portal, field OS, and AI governance.
  contact:
    name: Platform Team
  license:
    name: Proprietary
servers:
  - 
    url: "http://localhost:3000"
    description: Local development
  - 
    url: "http://localhost:8080"
    description: Docker compose
  - 
    url: "https://ceop.mirai-dx-platform.com"
    description: Production
...
paths:
...
  /api/v1/users:
    get:
      operationId: listUsers
      summary: Paginated list of users
      tags:
        - Users
      security:
        - 
          ApiKeyAuth: []
        - 
          BearerJwt: []
      parameters:
        - 
          $ref: "#/components/parameters/limitParam"
        - 
          $ref: "#/components/parameters/offsetParam"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - users
                  - count
                  - total
                  - limit
                  - offset
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
                  count:
                    type: integer
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createUser
      summary: Create a user
      tags:
        - Users
      security:
        - 
          ApiKeyAuth: []
        - 
          BearerJwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - organizationId
                - displayName
                - email
              properties:
                organizationId:
                  type: string
                displayName:
                  type: string
                email:
                  type: string
                  format: email
                status:
                  type: string
                  enum:
                    - invited
                    - active
                    - suspended
                    - deactivated
                roleIds:
                  type: array
                  items:
                    type: string
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/users/{id}:
    get:
      operationId: getUser
      summary: Get a user by ID
      tags:
        - Users
      security:
        - 
          ApiKeyAuth: []
        - 
          BearerJwt: []
      parameters:
        - 
          $ref: "#/components/parameters/idPath"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateUser
      summary: Update a user
      tags:
        - Users
      security:
        - 
          ApiKeyAuth: []
        - 
          BearerJwt: []
      parameters:
        - 
          $ref: "#/components/parameters/idPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName:
                  type: string
                status:
                  type: string
                  enum:
                    - invited
                    - active
                    - suspended
                    - deactivated
                roleIds:
                  type: array
                  items:
                    type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteUser
      summary: Soft-delete a user (status → deactivated)
      tags:
        - Users
      security:
        - 
          ApiKeyAuth: []
        - 
          BearerJwt: []
      parameters:
        - 
          $ref: "#/components/parameters/idPath"
      responses:
        204:
          description: Deactivated
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    User:
      type: object
      required:
        - id
        - organizationId
        - displayName
        - email
        - status
        - roleIds
        - createdAt
      properties:
        id:
          type: string
        organizationId:
          type: string
        displayName:
          type: string
        email:
          type: string
          format: email
        status:
          type: string
          enum:
            - invited
            - active
            - suspended
            - deactivated
        roleIds:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
...
  parameters:
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 200
      description: Number of items per page
    offsetParam:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
      description: Number of items to skip
    idPath:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource ID
...
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: "API key in the format `keyId:rawSecret`. The server validates via HMAC-SHA256 and constant-time comparison."
      bearerFormat: "keyId:rawSecret"
    BearerJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: HS256 JWT issued by POST /api/v1/auth/token. Include in Authorization header.